UPDATE: update authentication to cookies for access and refresh token | refactor auth

+++ src/main/java/com/deliveryplatform/auth/jwt/JwtConfig.java
@@ -9,6 +9,6 @@ import org.springframework.context.annotation.Configuration;
 @Data
 public class JwtConfig {
     private String secret;
-    private int accessTokenDuration;
-    private int refreshTokenDuration;
+    private long accessTokenDuration;
+    private long refreshTokenDuration;
 }
